[UIPQB-281] Implement fuzzy matching within 'Field' dropdown
